# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  id: number;
  playerName: string;
  score: number;
  wavesCleared: number;
  bossReached: number;
  isFullCompletion: boolean;
  createdAt: Date;
}